refactor: split overlay designation helpers into own part file (#3617)
province_sea_zone_detail_overlay_sections.dart exceeded the 700-line
game-widgets file-size gate (713 lines). Extract the town/capital
designation cluster (provinceOverlayIsCapital,
_capitalHolderDisplayNameForTile, provinceOverlayTileDesignationLine)
into province_sea_zone_detail_overlay_designation.dart as a new part
file. Behavior-preserving; sections file now 651 lines.

Refs #3617
