+++ app/lib/features/game/widgets/province_sea_zone_detail_overlay.dart
@@ -40,6 +40,7 @@ import 'province_overlay_unit_partition.dart';
 
 part 'province_sea_zone_detail_overlay_sections.dart';
 part 'province_sea_zone_detail_overlay_economic_military_sections.dart';
+part 'province_sea_zone_detail_overlay_designation.dart';
 
 class ProvinceSeaZoneDetailOverlay extends StatelessWidget {
   /// SPEC/ui/province-sea-zone-detail-overlay.md — [UiScreenIds.provinceSeaZoneOverlay].
